debian, setup.py: Add dependency on python3-tomli
Closes: #2169.

python3-coverage >= 6.0 requires python3-tomli to be installed when using
pyproject.toml. There is no hard dependency on tomli for the coverage package
perhaps due to its optional nature.

However, python3-toml is not available in Bullseye. So, require a version of
python3-coverage that does not require python3-tomli.

Tests:

- Run git-pbuilder for stable and unstable. Builds were successful.

- Run autopkgtest for stable and unstable. Dependencies were satisfied but there
was an unrelated error during test collection.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>