upgrades: Use n= in apt preferences
In the Debian bookworm-backports repository, the Suite has changed to
"stable-backports". The Codename is the more specific "bookworm-backports".

Bump upgrades app version to rewrite the apt preferences file.

Note: The app version is kept lower than the one in bookworm-backports
currently (17). This is to ensure that additional setup steps are applied when
the bookworm-backports package gets installed.

Helps: #2368.

Tests:

- `apt policy freedombox` shows priority 500 for package in bookworm-backports.

Signed-off-by: James Valleroy <jvalleroy@mailbox.org>
Reviewed-by: Sunil Mohan Adapa <sunil@medhas.org>